# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API for Testing",
    "description": "API desenvolvida por Jam Batista para treinamentos de QAs.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/login": {
      "post": {
        "tags": ["Login"],
        "summary": "Generate new Bearer token",
        "description": "Endpoint to generate a new token with default credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "default": "admin",
                    "description": "Nome de usuário para autenticação"
                  },
                  "password": {
                    "type": "string",
                    "default": "admin",
                    "description": "Senha para autenticação"
                  }
                },
                "required": ["username", "password"],
                "example": {
                  "username": "admin",
                  "password": "admin"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid credentials provided."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Authentication failed."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login-headers": {
      "post": {
        "summary": "Login usando headers",
        "description": "Realiza o login e valida se o header Authorization está presente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas ou header de autorização ausente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Credenciais inválidas"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Token de autorização no formato Bearer",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer <seu_token_aqui>"
            }
          }
        ]
      }
    },
    "/login-hard": {
      "post": {
        "tags": ["Login"],
        "summary": "Generate new Bearer token",
        "description": "Endpoint to generate a new token with default credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "default": "admin",
                    "description": "Nome de usuário para autenticação"
                  },
                  "password": {
                    "type": "string",
                    "default": "admin",
                    "description": "Senha para autenticação"
                  }
                },
                "required": ["username", "password"],
                "example": {
                  "username": "admin",
                  "password": "admin"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid credentials provided."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Authentication failed."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Returns a list of all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "description": "Creates a new user with any provided data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Any data can be sent to create a user",
                "example": {
                  "name": "John",
                  "last_name": "Doe",
                  "email": "john.doe@example.com",
                  "age": 30,
                  "address": "123 Main St"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get all projects",
        "description": "Retrieve a list of all projects",
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "description": "Creates a new project with the provided information",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Get all events",
        "description": "Retrieve a list of all events",
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event",
        "description": "Creates a new event with the provided information",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEvent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/approval": {
      "get": {
        "tags": ["Approval"],
        "summary": "Get all approvals",
        "description": "Retrieve a list of all user approvals.",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserApproval"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      },
      "post": {
        "tags": ["Approval"],
        "summary": "Create a new approval",
        "description": "Creates a new user approval with PENDING status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUserApproval"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User approval created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserApprovalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/validation/level1": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level2": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level3": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level4": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level5": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level6": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level7": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level8": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level9": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    },
    "/validation/level10": {
      "get": {
        "tags": ["challengers"],
        "summary": "Chellengers JSON",
        "description": "validation",
        "responses": {
          "200": {
            "description": "List of approvals",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No user approval data found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Website Redesign"
          },
          "leader": {
            "type": "string",
            "example": "Alice Johnson"
          },
          "description": {
            "type": "string",
            "example": "A complete redesign of the company website to improve user experience."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2024-01-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2024-06-01"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id_member": {
                  "type": "integer",
                  "example": 1
                },
                "member_name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "office": {
                  "type": "string",
                  "example": "Frontend Developer"
                }
              }
            }
          }
        }
      },
      "NewProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Website Redesign"
          },
          "leader": {
            "type": "string",
            "example": "Alice Johnson"
          },
          "description": {
            "type": "string",
            "example": "A complete redesign of the company website to improve user experience."
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2024-06-01"
          }
        },
        "required": ["name", "leader", "description", "endDate"]
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Tech Conference 2024"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-15"
          },
          "location": {
            "type": "string",
            "example": "New York"
          },
          "capacity": {
            "type": "integer",
            "example": 50
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "Alice Johnson"
                },
                "email": {
                  "type": "string",
                  "example": "alice.johnson@example.com"
                },
                "age": {
                  "type": "integer",
                  "example": 25
                }
              }
            }
          }
        }
      },
      "NewEvent": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Conference 2024"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-15"
          },
          "location": {
            "type": "string",
            "example": "New York"
          },
          "capacity": {
            "type": "integer",
            "example": 50
          }
        },
        "required": ["name", "date", "location", "capacity"]
      },
...
      "UserApproval": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "example": "12345678901"
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          }
        }
      },
      "NewUserApproval": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "example": "12345678901"
          }
        },
        "required": ["name", "email", "cpf"]
      },
      "UserApprovalResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "example": "12345678901"
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          }
        }
      },
...
    }
  }
}